# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/dashboard/stats": {
      "get": {
        "operationId": "DashboardController_stats",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 2,
              "pattern": "^[A-Za-z]{2}$",
              "type": "string"
            }
          },
          {
            "name": "crop",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "name": "harvestYear",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 10,
              "type": "string"
            }
          },
          {
            "name": "esgStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "APPROVED",
                "WARNING",
                "BLOCKED"
              ]
            }
          },
          {
            "name": "carStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "PENDING",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "minClimateRisk",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "maxClimateRisk",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shape completo (summary+analytics+ESG). Residual para clientes legados — first paint deve usar /summary."
          }
        },
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/dashboard/summary": {
      "get": {
        "operationId": "DashboardController_summary",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 2,
              "pattern": "^[A-Za-z]{2}$",
              "type": "string"
            }
          },
          {
            "name": "crop",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "name": "harvestYear",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 10,
              "type": "string"
            }
          },
          {
            "name": "esgStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "APPROVED",
                "WARNING",
                "BLOCKED"
              ]
            }
          },
          {
            "name": "carStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "PENDING",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "minClimateRisk",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "maxClimateRisk",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "First paint: KPIs, byState/byCrop/byLandUse/CAR/ESG. Caminho preferido do frontend."
          }
        },
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/dashboard/analytics": {
      "get": {
        "operationId": "DashboardController_analytics",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 2,
              "pattern": "^[A-Za-z]{2}$",
              "type": "string"
            }
          },
          {
            "name": "crop",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "name": "harvestYear",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 10,
              "type": "string"
            }
          },
          {
            "name": "esgStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "APPROVED",
                "WARNING",
                "BLOCKED"
              ]
            }
          },
          {
            "name": "carStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "PENDING",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "minClimateRisk",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "maxClimateRisk",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Secondary paint: séries (cropsByYear, farmsByMonth), topCities, risco climático por estado/cultura."
          }
        },
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/admin/revalidate/producers/{id}": {
      "post": {
        "operationId": "AdminRevalidateController_producer",
        "parameters": [
          {
            "name": "X-Admin-Token",
            "in": "header",
            "description": "Token administrativo (ADMIN_API_TOKEN)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revalida CNPJ do produtor via BrasilAPI",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevalidateResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/revalidate/farms/{id}": {
      "post": {
        "operationId": "AdminRevalidateController_farm",
        "parameters": [
          {
            "name": "X-Admin-Token",
            "in": "header",
            "description": "Token administrativo (ADMIN_API_TOKEN)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revalida cidade/UF da fazenda via BrasilAPI",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevalidateResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "Brain Agriculture API",
    "description": "API de gestão de produtores rurais, fazendas, safras e dashboard analítico. Sem autenticação no escopo atual — proteja a rede e use rate limit/CORS em produção.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RevalidateResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "previousStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ]
          },
          "newStatus": {
            "type": "string",
            "enum": [
              "VALIDATED",
              "PENDING_EXTERNAL_VALIDATION",
              "REJECTED"
            ]
          },
          "reason": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "previousStatus",
          "newStatus"
        ]
      }
    }
  }
}